Add user agent to API calls.
Unfortunately a unnamed third party has used this API, in their
product, and due to issues in their code, have generated a very
large amount of requests to our system. While we were debugging
this, a hint of the user agent would have been really useful.

This PR adds a default UA to it, Because this API can be used
in many applications, I've made it the __FILE__ var, since
that will give us the best chance of getting in contact with
a possibly broken script vendor.